+++ xdef/changelog.md
@@ -1,8 +1,10 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 41.0.4, release-date 2021-12-22
+# Version 41.0.4, release-date 2021-12-27
 * added new type validation methods `country` and `countries`
 * `<xd:json>` model in X-definition changed to `<xd:xon>`
+* fixed bug in datetime formationg when mask of a xdatetime contains `SSS`
+  and the value of milliseconds in datetime is zero
 * fixed bug NullpointerException in parsing of incorrect JSON data
 
 # Version 41.0.3, release-date 2021-12-05